+++ packages/theme_tailor/example/test/nullable_params_test.dart
@@ -1,0 +1,14 @@
+import 'package:example/nullable_params.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test(
+    'Nullable params are not required',
+    () {
+      const NullableParams(
+        anotherProperty: Colors.green,
+      );
+    },
+  );
+}